Put cleanup in a separate step to catch any and all error returns

+++ .github/scripts/manage-queue-state.sh
@@ -1,6 +1,9 @@
 #!/usr/bin/env bash
 
+# if --safely is passed we won't error if the merge branch doesn't exist
 remove_branch_from_queue () {
+  local safely="${1:-}"
+
   start_dir=$(pwd)
 
   cd $GITHUB_WORKSPACE/merge-queue-state
@@ -14,10 +17,6 @@ remove_branch_from_queue () {
         )
 
   if [ -z "$state" ]; then
-    # can't call sad_ending from here as that calls this and we could be stuck
-    # in a loop
-    sad_update "💣 Error - failed updating the merge queue, aborting"
-    unlock_merge_queue --force
     exit 1
   fi
 
@@ -26,6 +25,11 @@ remove_branch_from_queue () {
   unlock_merge_queue
 
   cd $GITHUB_WORKSPACE/project
+  # if we're running --safely return if the branch doesn't exist so we don't error
+  if [[ "$safely" == "--safely" && ! $(git branch -r | grep "$MERGE_BRANCH") ]]; then
+    cd $start_dir
+    return 0
+  fi
   git push --delete origin $MERGE_BRANCH
 
   cd $start_dir
@@ -45,10 +49,6 @@ remove_descendants_from_queue () {
         )
 
   if [ -z "$state" ]; then
-    # can't call sad_ending from here as that calls this and we could be stuck
-    # in a loop
-    sad_update "💣 Error - failed updating the merge queue, aborting"
-    unlock_merge_queue --force
     exit 1
   fi
